IMP: Complete CI Pipeline and fix build

+++ src/Beng.Specta.Compta.Infrastructure/Data/EfRepository.cs
@@ -15,7 +15,7 @@ namespace Beng.Specta.Compta.Infrastructure.Data
             _dbContext = dbContext;
         }
 
-        public T GetById<T>(int id) where T : BaseEntity
+        public T GetById<T>(long id) where T : BaseEntity
         {
             return _dbContext.Set<T>().SingleOrDefault(e => e.Id == id);
         }
