+++ src/Beng.Specta.Compta.SharedKernel/Interfaces/IRepository.cs
@@ -4,7 +4,7 @@ namespace Beng.Specta.Compta.SharedKernel.Interfaces
 {
     public interface IRepository
     {
-        T GetById<T>(int id) where T : BaseEntity;
+        T GetById<T>(long id) where T : BaseEntity;
         List<T> List<T>() where T : BaseEntity;
         T Add<T>(T entity) where T : BaseEntity;
         void Update<T>(T entity) where T : BaseEntity;
